+++ src/ebusd/mainloop.cpp
@@ -108,7 +108,7 @@ MainLoop::MainLoop(const struct options& opt, BusHandler* busHandler,
   : Thread(), m_busHandler(busHandler), m_protocol(busHandler->getProtocol()), m_reconnectCount(0),
     m_userList(opt.accessLevel), m_messages(messages),
     m_scanHelper(scanHelper), m_address(opt.address), m_scanConfig(opt.scanConfig),
-    m_initialScan(opt.readOnly ? ESC : opt.initialScan), m_scanRetries(opt.scanRetries),
+    m_initialScan(opt.readOnly ? (symbol_t)ESC : opt.initialScan), m_scanRetries(opt.scanRetries),
     m_scanStatus(SCAN_STATUS_NONE), m_polling(opt.pollInterval > 0), m_enableHex(opt.enableHex),
     m_shutdown(false), m_runUpdateCheck(opt.updateCheck), m_httpClient(), m_requestQueue(requestQueue) {
   if (opt.aclFile[0]) {
@@ -352,8 +352,9 @@ void MainLoop::run() {
       m_messages->lock();
       m_messages->findAll("", "", "*", false, true, true, true, true, true, sinkSince, now, false, &messages);
       for (const auto message : messages) {
+        bool changed = message->getLastChangeTime() >= sinkSince;
         for (const auto dataSink : dataSinks) {
-          dataSink->notifyUpdate(message);
+          dataSink->notifyUpdate(message, changed);
         }
       }
       m_messages->unlock();
@@ -480,6 +481,13 @@ result_t MainLoop::decodeRequest(Request* req, bool* connected, RequestMode* req
   if (cmd == "INJECT") {
     if (m_enableHex) {
       return executeInject(args, ostream);
+    }
+    *ostream << "ERR: command not enabled";
+    return RESULT_OK;
+  }
+  if (cmd == "ANSWER") {
+    if (m_enableHex && m_protocol->isAnswering()) {
+      return executeAnswer(args, ostream);
     }
     *ostream << "ERR: command not enabled";
     return RESULT_OK;
@@ -668,7 +676,7 @@ result_t MainLoop::executeRead(const vector<string>& args, const string& levels,
       if (dest) {
         dstAddress = address;
       } else {
-        srcAddress = address == m_address ? SYN : address;
+        srcAddress = address == m_address ? (symbol_t)SYN : address;
       }
     } else if (args[argPos] == "-p") {
       argPos++;
@@ -925,7 +933,7 @@ result_t MainLoop::executeWrite(const vector<string>& args, const string levels,
       if (dest) {
         dstAddress = address;
       } else {
-        srcAddress = address == m_address ? SYN : address;
+        srcAddress = address == m_address ? (symbol_t)SYN : address;
       }
     } else if (args[argPos] == "-c") {
       argPos++;
@@ -1102,7 +1110,7 @@ result_t MainLoop::parseHexAndSend(const vector<string>& args, size_t& argPos, b
       if (ret != RESULT_OK || !isValidAddress(address, false) || !isMaster(address)) {
         return RESULT_ERR_INVALID_ADDR;
       }
-      srcAddress = address == m_address ? SYN : address;
+      srcAddress = address == m_address ? (symbol_t)SYN : address;
     } else if (args[argPos] == "-n") {
       autoLength = true;
     } else {
@@ -1170,7 +1178,7 @@ result_t MainLoop::executeInject(const vector<string>& args, ostringstream* ostr
     if (!m_scanHelper->parseMessage(args[argPos++], false, &master, &slave)) {
       return RESULT_ERR_INVALID_ARG;
     }
-    m_busHandler->notifyProtocolMessage(false, master, slave);
+    m_busHandler->notifyProtocolMessage(md_recv, master, slave);
     return RESULT_OK;
   }
   *ostream << "usage: inject QQZZPBSBNN[DD]*/[NN[DD]*]\n"
@@ -1180,6 +1188,83 @@ result_t MainLoop::executeInject(const vector<string>& args, ostringstream* ostr
               "  PB SB  primary/secondary command byte\n"
               "  NN     number of following data bytes\n"
               "  DD     data byte(s)";
+  return RESULT_OK;
+}
+
+result_t MainLoop::executeAnswer(const vector<string>& args, ostringstream* ostream) {
+  size_t argPos = 1;
+  symbol_t srcAddress = SYN;
+  symbol_t dstAddress = SYN;
+  bool master = false;
+  while (args.size() > argPos && args[argPos][0] == '-') {
+    if (args[argPos] == "-s" && argPos + 1 < args.size()) {
+      result_t ret;
+      argPos++;
+      symbol_t address = (symbol_t)parseInt(args[argPos].c_str(), 16, 0, 0xff, &ret);
+      if (ret != RESULT_OK || !isValidAddress(address, false) || !isMaster(address)) {
+        return RESULT_ERR_INVALID_ADDR;
+      }
+      srcAddress = address;
+    } else if (args[argPos] == "-d" && argPos + 1 < args.size()) {
+      result_t ret;
+      argPos++;
+      symbol_t address = (symbol_t)parseInt(args[argPos].c_str(), 16, 0, 0xff, &ret);
+      if (ret != RESULT_OK || !isValidAddress(address)) {
+        return RESULT_ERR_INVALID_ADDR;
+      }
+      dstAddress = address;
+    } else if (args[argPos] == "-m") {
+      master = true;
+    } else {
+      argPos = 0;  // print usage
+      break;
+    }
+    argPos++;
+  }
+  MasterSymbolString id;
+  if (argPos > 0 && argPos < args.size()) {
+    result_t ret = id.parseHex(args[argPos++]);
+    if (ret != RESULT_OK) {
+      return ret;
+    }
+    if (id.size() < 2 || id.size() > 6) {
+      return RESULT_ERR_INVALID_POS;
+    }
+  }
+  SlaveSymbolString answer;
+  answer.push_back(0);  // room for length byte
+  if (argPos > 0 && argPos < args.size()) {
+    result_t ret = answer.parseHex(args[argPos++]);
+    if (ret != RESULT_OK) {
+      return ret;
+    }
+    if (answer.size() > 16) {
+      return RESULT_ERR_INVALID_POS;
+    }
+  }
+  answer.adjustHeader();
+  if (argPos < args.size()) {
+    argPos = 0;  // print usage
+  }
+  if (argPos <= 1) {
+    *ostream << "usage: answer [-m] [-s QQ] [-d ZZ] PBSB[ID]* [DD]*\n"
+                " Answer to a message from the bus.\n"
+                "  -m     destination is a master\n"
+                "  -s QQ  source address to limit to\n"
+                "  -d ZZ  override destination address (instead of own address)\n"
+                "  PB SB  primary/secondary command byte\n"
+                "  ID     further ID bytes\n"
+                "  DD     data bytes (only length used with -m)";
+    return RESULT_OK;
+  }
+  if (isMaster(dstAddress)) {
+    master = true;
+  } else if (dstAddress == SYN) {
+    dstAddress = master ? m_address : getSlaveAddress(m_address);
+  }
+  if (!m_protocol->setAnswer(srcAddress, dstAddress, id[0], id[1], id.data()+2, id.size()-2, answer)) {
+    return RESULT_ERR_INVALID_ARG;
+  }
   return RESULT_OK;
 }
 
@@ -1915,12 +2000,13 @@ result_t MainLoop::executeHelp(ostringstream* ostream) {
       " listen|l  Listen for updates:    listen [-v|-V] [-n|-N] [-u|-U] [stop]\n"
       " hex       Send hex data:         hex [-s QQ] [-n] ZZPBSB[NN][DD]* (if enabled)\n"
       " inject    Inject hex data:       inject QQZZPBSBNN[DD]*/[NN[DD]*] (if enabled)\n"
+      " answer    Answer a message:      answer [-m] [-s QQ] [-d ZZ] PBSB[ID]* [DD]* (if enabled)\n"
       " direct    Enter direct mode\n"
       " state|s   Report bus state\n"
       " info|i    Report information about the daemon, configuration, seen participants, and the device.\n"
       " grab|g    Grab messages:         grab [stop]\n"
       "           Report the messages:   grab result [all|decode]\n"
-      " define    Define new message:    define [-r] DEFINITION\n"
+      " define    Define new message:    define [-r] DEFINITION (if enabled)\n"
       " decode|d  Decode field(s):       decode [-v|-V] [-n|-N] DEFINITION DD[DD]*\n"
       " encode|e  Encode field(s):       encode DEFINITION VALUE[;VALUE]*\n"
       " scan      Scan slaves:           scan [full|ZZ]\n"